requirements.txt
# Core dependencies for AI Employee System
anthropic>=0.3.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
playwright>=1.40.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
watchdog>=3.0.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
cryptography>=41.0.8
pyyaml>=6.0.1
aiofiles>=23.2.1
asyncio-mqtt>=0.13.0
redis>=5.0.1
celery>=5.3.4
alembic>=1.13.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-slugify>=8.0.1
phonenumbers>=8.13.21
twilio>=8.10.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
openpyxl>=3.1.2
pandas>=2.1.4
numpy>=1.26.2
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0
scikit-learn>=1.3.2
nltk>=3.8.1
spacy>=3.7.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.22
langchain>=0.0.335
langchain-community>=0.0.19
langchain-openai>=0.0.5
tiktoken>=0.5.2
python-telegram-bot>=20.7
selenium>=4.15.2
webdriver-manager>=4.0.1
pyautogui>=0.9.54
keyboard>=0.13.5
pytesseract>=0.3.10
opencv-python>=4.8.1.78
pillow>=10.1.0
pdf2image>=1.16.3
pyttsx3>=2.90
SpeechRecognition>=4.0.1
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.13
cachetools>=5.3.2
oauth2client>=4.1.3
gspread>=5.12.0
pymongo>=4.6.0
motor>=3.3.2
aioredis>=2.0.1
aiohttp>=3.9.0
httpx>=0.25.2
httpx-sse>=0.4.0
sse-starlette>=1.6.5
websockets>=12.0
starlette>=0.27.0
anyio>=3.7.1
typing-extensions>=4.8.0
idna>=3.4
certifi>=2023.11.17
charset-normalizer>=3.3.2
urllib3>=2.1.0
h11>=0.14.0
sniffio>=1.3.0
click>=8.1.7
colorama>=0.4.6
pydantic-core>=2.14.5
email-validator>=2.1.0
jinja2>=3.1.2
MarkupSafe>=2.1.3
orjson>=3.9.10
python-dateutil>=2.8.2
pytz>=2023.3.post1
tzdata>=2023.3
zipp>=3.17.0
iniconfig>=2.0.0
packaging>=23.2
pluggy>=1.4.0
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.1
pre-commit>=3.5.0